pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "fermilatprogram"
version = "1.0.0"
description = "Fermi-LAT GRB数据分析工具包"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "GRB Analysis Team", email = "grb@example.com"},
]
maintainers = [
    {name = "GRB Analysis Team", email = "grb@example.com"},
]
keywords = ["fermi-lat", "grb", "gamma-ray-burst", "astronomy", "astrophysics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Astronomy",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy>=1.19.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "astropy>=4.0",
    "fermipy>=1.0.0",
    "matplotlib>=3.3.0",
    "seaborn>=0.11.0",
    "PyYAML>=5.4.0",
    "h5py>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.800",
    "pre-commit>=2.15.0",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=0.5",
    "myst-parser>=0.15.0",
]
test = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "pytest-xdist>=2.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/fermilatprogram"
Documentation = "https://fermilatprogram.readthedocs.io/"
Repository = "https://github.com/yourusername/fermilatprogram.git"
"Bug Tracker" = "https://github.com/yourusername/fermilatprogram/issues"
Changelog = "https://github.com/yourusername/fermilatprogram/blob/main/CHANGELOG.md"

[project.scripts]
grb-analyze = "fermilatprogram.lkmulty:main"
grb-download = "fermilatprogram.download:main"
grb-config = "fermilatprogram.Generate_gconfig:main"

[tool.setuptools]
packages = ["fermilatprogram"]

[tool.setuptools.package-data]
fermilatprogram = ["*.yaml", "*.yml", "*.txt", "*.md"]

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310']
include = '\.pyi?$'
extend-exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["fermilatprogram"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "fermipy.*",
    "astropy.*",
    "matplotlib.*",
    "seaborn.*",
]
ignore_missing_imports = true